Add files via upload
Sentiment analysis plays a crucial role in understanding public opinion and sentiment on social media platforms. This project aims to perform sentiment analysis on a large dataset of Twitter posts to gain insights into user sentiments regarding a specific topic. The objective of this project is to classify tweets sentiment as positive or negative sentiments. For suppose someone has tweeted “Disappointed with the governments by their specific policy or action ” this on the social media which indicated it’s a negative comment on the current government of India. We aim to identify prevalent sentiments and track their changes over time. Which will eventually reduce the rate of irrelevant tweets, so that the user will experience a user friendly environment. We employed the Natural Language Toolkit (NLTK) and scikit-learn libraries(sk-learn) in Python for text preprocessing and feature extraction. A Deep learning model, based on Recurrent Neural Network Module (RNN) long-short term memory (LSTM), trained and evaluated using a dataset of 1,000 labelled tweets as a prototype and a final module trained on 16,00,000 labelled tweets. Our sentiment analysis model achieved a Training Accuracy of 84.5% and a Test Accuracy of 73.9% on the prototype model and the Final models has a Training Accuracy of 94.7% and a Test Accuracy of 95.4%, the test dataset. Implications and Applications: The sentiment analysis model can be applied to various domains for real-time public opinion tracking and social media monitoring.